Fixed setting viewport in multiplayer games.
It was broken in commit 5c4834cf31bf22bad306f4820d2693b880a5aec7.

We mix irr_driver->getVideoDriver()->setViewPort() and glViewport(...) functions and thus the first one, which is executed during camera activation, in some cases doesn't work. It compares the viewport to values which has been set last time and thinks that nothing changed.

It only happens on single player (with single camera), so we just restore the viewport to fullscreen after scene rendering to workaround the issue.

Actually the problem was that it was set after camera activation, which overwrote proper values.

I also fixed SSAO, RSM and normals vizualization in multiplayer games.