Expose getKartType(int kartId) to scripting
This change attempts to expose the KartType (local player, networked player, AI, etc.) to scripting.
This is useful if a track creator wants an action to only occur if triggered by a real player, or only if triggered by a non-networked player (e.g. alternate route music in Frozen Drive unofficial add-on).

I have rushed this in to try and get it added in STK 1.1 and I have not yet been able to configure and build STK from source, so this change is currently untested!

I have placed this in the Track section because it uses the race manager. Please let me know if the Kart section is more appropriate.